Feat: enable user customization via setup function
Also moved cmp source related code into a separate file.
