+++ after/plugin/cmp_nvim_ultisnips.lua
@@ -1,1 +1,1 @@
-require'cmp'.register_source('ultisnips', require'cmp_nvim_ultisnips'.new())
+require('cmp').register_source('ultisnips', require('cmp_nvim_ultisnips').create_source())
